# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIMarket",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/shops": {
      "post": {
        "tags": [
          "shops"
        ],
        "summary": "Create Shop",
        "description": "[Seller/Admin]",
        "operationId": "create_shop_api_v1_shops_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/shops/me": {
      "get": {
        "tags": [
          "shops"
        ],
        "summary": "Get My Shops",
        "operationId": "get_my_shops_api_v1_shops_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/shops/{shop_id}": {
      "patch": {
        "tags": [
          "shops"
        ],
        "summary": "Update Shop",
        "description": "[Seller/Admin]",
        "operationId": "update_shop_api_v1_shops__shop_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "shops"
        ],
        "summary": "Delete Shop",
        "description": "[Seller/Admin]",
        "operationId": "delete_shop_api_v1_shops__shop_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "shops"
        ],
        "summary": "Get Shop",
        "operationId": "get_shop_api_v1_shops__shop_id__get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shops/{shop_id}/favourite": {
      "post": {
        "tags": [
          "shops"
        ],
        "summary": "User Change Favourite Status Shop",
        "description": "[Buyer User]",
        "operationId": "user_change_favourite_status_shop_api_v1_shops__shop_id__favourite_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "shop_id": {
            "type": "integer",
            "title": "Shop Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price",
            "examples": [
              "199.99"
            ]
          },
          "discount_percent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Discount Percent",
            "examples": [
              "10.5"
            ]
          },
          "final_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Final Price",
            "examples": [
              "199.99"
            ]
          },
          "rating": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating",
            "examples": [
              4.8
            ]
          },
          "reviews_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reviews Count"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shop_id",
          "title",
          "images",
          "price",
          "discount_percent",
          "final_price",
          "reviews_count",
          "quantity"
        ],
        "title": "ProductCard"
      },
...
      "ShopCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "title": "Title"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ShopCreateRequest"
      },
      "ShopCreateResponse": {
        "properties": {
          "shop": {
            "$ref": "#/components/schemas/ShopResponse"
          }
        },
        "type": "object",
        "required": [
          "shop"
        ],
        "title": "ShopCreateResponse"
      },
      "ShopDeleteResponse": {
        "properties": {
          "shop_id": {
            "type": "integer",
            "title": "Shop Id"
          }
        },
        "type": "object",
        "required": [
          "shop_id"
        ],
        "title": "ShopDeleteResponse"
      },
      "ShopResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "favourites_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Favourites Count"
          },
          "reviews_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reviews Count"
          },
          "rating": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rating"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductCard"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "favourites_count",
          "reviews_count",
          "rating",
          "products"
        ],
        "title": "ShopResponse"
      },
      "ShopSummaryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "reviews_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reviews Count"
          },
          "rating": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rating"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "reviews_count",
          "rating",
          "is_active"
        ],
        "title": "ShopSummaryResponse"
      },
      "ShopUpdateRequest": {
        "properties": {
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ShopUpdateRequest"
      },
      "ShopUpdateResponse": {
        "properties": {
          "shop": {
            "$ref": "#/components/schemas/ShopResponse"
          }
        },
        "type": "object",
        "required": [
          "shop"
        ],
        "title": "ShopUpdateResponse"
      },
      "ShopsResponse": {
        "properties": {
          "shops": {
            "items": {
              "$ref": "#/components/schemas/ShopSummaryResponse"
            },
            "type": "array",
            "title": "Shops"
          }
        },
        "type": "object",
        "required": [
          "shops"
        ],
        "title": "ShopsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}